♻️ refactor(api): update import paths to use models directory
