+++ apps/api/src/application/use-cases/auth/validate-session.usecase.ts
@@ -1,7 +1,7 @@
 import { sessionExpiresSpan, sessionRefreshSpan } from "../../../common/constants";
-import { Session } from "../../../entities/session";
 import type { ISessionRepository } from "../../../interface-adapter/repositories/session";
 import type { IUserRepository } from "../../../interface-adapter/repositories/user";
+import { Session } from "../../../models/entities/session";
 import type { ISessionTokenService } from "../../../services/session-token";
 import type { IValidateSessionUseCase, IValidateSessionUseCaseResult } from "./interface/validate-session.usecase";
