+++ apps/api/src/application/use-cases/email-verification/email-verification-request.usecase.ts
@@ -1,8 +1,8 @@
 import { emailVerificationCodeExpiresSpan } from "../../../common/constants";
 import { generateRandomString, ulid } from "../../../common/utils";
-import type { User } from "../../../entities/user";
 import type { IEmailVerificationCodeRepository } from "../../../interface-adapter/repositories/email-verification-code";
 import type { IUserRepository } from "../../../interface-adapter/repositories/user";
+import type { User } from "../../../models/entities/user";
 import type {
 	IEmailVerificationRequestUseCase,
 	IEmailVerificationRequestUseCaseResult,
