+++ apps/api/src/application/use-cases/email-verification/interface/email-verification-confirm.usecase.interface.ts
@@ -1,4 +1,4 @@
-import type { User } from "../../../../entities/user";
+import type { User } from "../../../../models/entities/user";
 
 export interface IEmailVerificationConfirmUseCaseResult {
 	success: boolean;
