+++ apps/api/src/application/use-cases/oauth/oauth-login-callback.usecase.ts
@@ -1,8 +1,8 @@
 import { sessionExpiresSpan } from "../../../common/constants";
-import type { OAuthProvider } from "../../../entities/oauth-account";
 import type { IOAuthProviderGateway } from "../../../interface-adapter/gateway/oauth-provider";
 import type { IOAuthAccountRepository } from "../../../interface-adapter/repositories/oauth-account";
 import type { ISessionRepository } from "../../../interface-adapter/repositories/session";
+import type { OAuthProvider } from "../../../models/entities/oauth-account";
 import type { ISessionTokenService } from "../../../services/session-token";
 import type {
 	IOAuthLoginCallbackUseCase,
