+++ apps/api/src/application/use-cases/user/change-email.usecase.ts
@@ -1,6 +1,6 @@
-import type { User } from "../../../entities/user";
 import type { IEmailVerificationCodeRepository } from "../../../interface-adapter/repositories/email-verification-code";
 import type { IUserRepository } from "../../../interface-adapter/repositories/user";
+import type { User } from "../../../models/entities/user";
 import type { IChangeEmailUseCase, IChangeEmailUseCaseResult } from "./interfaces/change-email.usecase.interface";
 
 export class ChangeEmailUseCase implements IChangeEmailUseCase {
