+++ apps/api/src/interface-adapter/gateway/oauth-provider/utils/select-oauth-provider-gateway.ts
@@ -1,4 +1,4 @@
-import type { OAuthProvider } from "../../../../entities/oauth-account";
+import type { OAuthProvider } from "../../../../models/entities/oauth-account";
 import type { OAuthProviderEnv } from "../../../../modules/env";
 import type { IOAuthProviderGateway } from "../interface/oauth-provider.gateway.interface";
 import { DiscordOAuthGateway } from "../providers/discord.gateway";
