+++ apps/api/src/interface-adapter/presenter/user.presenter.ts
@@ -1,5 +1,5 @@
 import { t } from "elysia";
-import type { User } from "../../entities/user";
+import type { User } from "../../models/entities/user";
 
 export type UserPresenterResult = {
 	id: string;
