+++ apps/api/src/interface-adapter/repositories/email-verification-code/interface/email-verification-code.repository.interface.ts
@@ -1,4 +1,4 @@
-import type { EmailVerificationCode } from "../../../../entities/email-verification-code";
+import type { EmailVerificationCode } from "../../../../models/entities/email-verification-code";
 
 export interface IEmailVerificationCodeRepository {
 	findByUserId(userId: EmailVerificationCode["userId"]): Promise<EmailVerificationCode | null>;
