+++ apps/api/src/interface-adapter/repositories/user-credential/interface/user-credential.repository.interface.ts
@@ -1,4 +1,4 @@
-import type { UserCredential } from "../../../../entities/user-credential";
+import type { UserCredential } from "../../../../models/entities/user-credential";
 
 export interface IUserCredentialRepository {
 	create(
