+++ apps/api/src/interface-adapter/repositories/user-credential/user-credential.repository.ts
@@ -1,6 +1,6 @@
 import { eq } from "drizzle-orm";
-import { UserCredential } from "../../../entities/user-credential";
 import type { DrizzleService } from "../../../infrastructure/drizzle";
+import { UserCredential } from "../../../models/entities/user-credential";
 import type { IUserCredentialRepository } from "./interface/user-credential.repository.interface";
 
 export class UserCredentialRepository implements IUserCredentialRepository {
