+++ apps/api/src/interface-adapter/repositories/user/tests/find-by-session-id.test.ts
@@ -1,7 +1,7 @@
 import { env } from "cloudflare:test";
 import { beforeAll, describe, expect, test } from "vitest";
-import { User } from "../../../../entities/user";
 import { DrizzleService } from "../../../../infrastructure/drizzle";
+import { User } from "../../../../models/entities/user";
 import { SessionTableHelper, UserTableHelper } from "../../../../tests/helpers";
 import { UserRepository } from "../user.repository";
