+++ apps/api/src/modules/auth-guard/auth-guard.plugin.ts
@@ -1,11 +1,11 @@
 import { ValidateSessionUseCase } from "../../application/use-cases/auth";
 import { SESSION_COOKIE_NAME } from "../../common/constants";
 import { readBearerToken } from "../../common/utils";
-import type { Session } from "../../entities/session";
-import type { User } from "../../entities/user";
 import { DrizzleService } from "../../infrastructure/drizzle";
 import { SessionRepository } from "../../interface-adapter/repositories/session";
 import { UserRepository } from "../../interface-adapter/repositories/user";
+import type { Session } from "../../models/entities/session";
+import type { User } from "../../models/entities/user";
 import { SessionTokenService } from "../../services/session-token";
 import { ElysiaWithEnv } from "../elysia-with-env";
 import { UnauthorizedException } from "../error/exceptions";
